Google Map w/ Certified API included - NO GUI **ROUGH**
I had to redo the maps portion to include the Google certified API so we
can get user data yaddyyadda.

When I changed the API the only thing that was giving me issues was the
onScreenOrientation in the manifest, which I changed to "Portrait"
instead of "locked" for the time being so I could get something up to
test.

ALSO changed the MinSDK to 9 in the manifest to work with some of the
libraries and such. Was not able to get past the login screen though to
test the MAP on device and sample GPS.

More updates later